chore: adjust sizing to fit two-per-row on small screens
